Add GraphQL as a Service
