+++ aptiche/lib/utils/string.dart
@@ -1,8 +1,13 @@
-//SVG Strings
+abstract class Strings {
+  //SVG Strings
 
-const String loginScreenSVG = 'assets/svgs/login_screen.svg';
-const String dataEntryScreen = 'assets/svgs/data_entry_screen.svg';
-const String homeScreenSVG = 'assets/svgs/TakeAQuiz.svg';
+  static const String loginScreenSVG = 'assets/svgs/login_screen.svg';
+  static const String dataEntryScreen = 'assets/svgs/data_entry_screen.svg';
+  static const String homeScreenSVG = 'assets/svgs/TakeAQuiz.svg';
 
 //PNG Strings
-const String quizPNG = 'assets/images/quiz.png';
+  static const String quizPNG = 'assets/images/quiz.png';
+
+  static const String GRAPHQL_URL =
+      'https://aptiche.dscnitrourkela.org/graphql?apikey=SriramForTheDiro';
+}
