+++ aptiche/lib/views/login/logincontroller.dart
@@ -22,6 +22,7 @@ class LoginController extends GetxController {
       smsCode.toString(),
       verificationId.toString(),
     );
+
     loading.value = false;
   }
