+++ aptiche/pubspec.yaml
@@ -19,6 +19,7 @@ dependencies:
   font_awesome_flutter: ^9.1.0
   get: ^4.1.4
   url_launcher: ^6.0.7
+  graphql: ^5.0.0
 
 dev_dependencies:
   flutter_test:
